fix reflective description of EParam. This fixes favour overflow, favour was not overflowing it was just being truncated by being cast through int16. Note that this also causes favour level maximum to be 9 instead of 6, which affects balance. Note that this requires regenerating generated source fiels including inc/dispatch.h

+++ inc/Api.h
@@ -685,7 +685,7 @@ system int16   T_EVENTINFO::Event;
 system int16   T_EVENTINFO::EXVal;
 system int16   T_EVENTINFO::EYVal;
 system int16   T_EVENTINFO::ESpellNum;
-system int16   T_EVENTINFO::EParam;
+system int32   T_EVENTINFO::EParam;
 system int16   T_EVENTINFO::EDir;
 system int8    T_EVENTINFO::vRoll;
 system int8    T_EVENTINFO::AType;
